Revert "New transaction id approach with auto increment numeric values"
This reverts commit 754aeb2c271e9edfed6c911afcfce830e5d8b5b6.

+++ src/tainted/transaction.h
@@ -26,7 +26,7 @@ using SharedRangesPool = iast::container::QueuedPool<SharedRanges, iast::Limits:
 
 namespace iast {
 namespace tainted {
-using transaction_key_t = uint64_t;
+using transaction_key_t = uintptr_t;
 
 class Transaction {
  public:
